+++ server/personaje.h
@@ -5,19 +5,19 @@ class Personaje{
     private:
         uint8_t type;
         uint8_t id;
-        uint16_t x_pos;
-        uint16_t y_pos;
+        float x_pos;
+        float y_pos;
         uint8_t orientation;
 
     public:
         Personaje();
         void move(uint8_t movement);
-        uint16_t getXPos();
-        uint16_t getYPos();
+        float getXPos();
+        float getYPos();
         uint8_t getOrientation();
         uint8_t getId();
-        void setXPos(uint16_t x_pos, bool isRight);
-        void setYPos(uint16_t y_pos, bool isJump);
+        void setXPos(float x_pos, bool isRight);
+        void setYPos(float y_pos, bool isJump);
         void setOrientation(uint8_t orientation);
         void setId(uint8_t id);
